Add Item ID generation function Rewrite the tests

+++ pallets/nac-managing/src/mock.rs
@@ -90,7 +90,7 @@ parameter_types! {
 }
 
 type CollectionId = u32;
-type ItemId = u32;
+type ItemId = u64;
 
 impl pallet_uniques::Config for Test {
     type RuntimeEvent = RuntimeEvent;
